% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_click}
\alias{add_click}
\title{Retrieve click in Shiny}
\usage{
add_click(map, inputId, layerId = NULL, action = "click")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{layerId}{Name of a variable present in data to filter results returned,
if \code{NULL} (default) all columns are returned.}

\item{action}{What triggers input value server-side, \code{click} or \code{dblclick}.}
}
\description{
Retrieve click in Shiny
}
\examples{
\dontrun{
library( r2d3maps )
library( r2d3 )
library( rnaturalearth )
library( shiny )

france <- ne_states(country = "france", returnclass = "sf")
france <- france[france$type_en \%in\% "Metropolitan department", ]

if (interactive()) {
  ui <- fluidPage(

    fluidRow(
      column(
        offset = 2, width = 8,
        tags$h2("r2d3maps Shiny example"),
        fluidRow(
          column(
            width = 6,
            d3Output(outputId = "map")
          ),
          column(
            width = 6,
            verbatimTextOutput(outputId = "res_click")
          )
        )
      )
    )

  )

  server <- function(input, output, session) {

    output$map <- renderD3({
      d3_map(shape = france) \%>\%
        add_tooltip() \%>\%
        add_click(
          inputId = "myclick",
          layerId = "name", # return only name,
          # NULL to get all data
          action = "dblclick" # on double click,
          # use "click" for simple click
        )
    })

    output$res_click <- renderPrint({
      str(input$myclick, max.level = 2)
    })
  }

  shinyApp(ui, server)
}
}
}
