% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.R
\name{add_continuous_breaks}
\alias{add_continuous_breaks}
\title{Add continuous scale to a map}
\usage{
add_continuous_breaks(map, var, palette = "viridis", direction = 1,
  n_breaks = 5, style = "pretty")
}
\arguments{
\item{map}{A \code{d3_map} \code{htmlwidget} object.}

\item{var}{Variable to map.}

\item{palette}{Color palette, you can use Viridis or Brewer color palette.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{n_breaks}{Number of breaks to cut data (depending on \code{style}, number of breaks can be re-computed).}

\item{style}{Style for computing breaks, see \code{\link[classInt]{classIntervals}}.}
}
\description{
Map a continuous numerical variables by
cutting it into class intervals.
}
\examples{
library( r2d3maps )
library( rnaturalearth )

# data
tunisia <- ne_states(country = "tunisia", returnclass = "sf")

# fake percentage
tunisia$p <- sample.int(100, nrow(tunisia))

# fake continuous var
tunisia$foo <- sample.int(1e5, nrow(tunisia))


# Tunisia
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "p")

# different color palette
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "p", palette = "Greens")

# legend
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "p",
                       palette = "inferno",
                       direction = -1) \%>\%
  add_legend(title = "Percentage", suffix = "\%")



# different style of breaks

# equal
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "foo",
                       palette = "inferno",
                       direction = -1,
                       style = "equal") \%>\%
  add_legend(title = "foo", d3_format = ".0f")

# quantile
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "foo",
                       palette = "inferno",
                       direction = -1,
                       style = "quantile") \%>\%
  add_legend(title = "foo", d3_format = ".0f")

# pretty
d3_map(shape = tunisia) \%>\%
  add_continuous_breaks(var = "foo",
                       palette = "inferno",
                       direction = -1,
                       style = "pretty") \%>\%
  add_legend(title = "foo", d3_format = ".0f")
}
