% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_discrete_scale}
\alias{add_discrete_scale}
\title{Title}
\usage{
add_discrete_scale(map, var, palette = "viridis", direction = 1,
  na.color = "#D8D8D8")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{var}{Variable to map}

\item{palette}{Color palette, you can use Viridis or Brewer color palette.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{na.color}{Color to use for missing values.}
}
\description{
Title
}
\examples{
# todo
}
