% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_labs}
\alias{add_labs}
\title{Labs for a r2d3 map}
\usage{
add_labs(map, title = NULL)
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{title}{The map title.}
}
\value{
A \code{r2d3map} \code{htmlwidget} object.
}
\description{
Labs for a r2d3 map
}
\examples{
\dontrun{

# todo

}
}
