% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-select.R
\name{add_select_input}
\alias{add_select_input}
\title{Add a select menu above a map}
\usage{
add_select_input(map, label = NULL, choices)
}
\arguments{
\item{map}{A \code{d3_cartogram} \code{htmlwidget} object.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from. Can be a named list. Values must be variables names.}
}
\description{
Use this to update variable used in \code{d3_cartogram}.
}
\examples{
# todo
}
