% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_cartogram.R
\name{d3_cartogram}
\alias{d3_cartogram}
\title{Create a cartogram in D3}
\usage{
d3_cartogram(shape, projection = "Mercator", stroke_col = "#fff",
  stroke_width = 0.5, width = NULL, height = NULL)
}
\arguments{
\item{shape}{A \code{sf} or \code{sp} object.}

\item{projection}{D3 projection to use.}

\item{stroke_col}{Color of shape contours.}

\item{stroke_width}{Width of shape contour in pixels.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Create a cartogram in D3
}
\examples{
# todo
}
