% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{d3_map}
\alias{d3_map}
\title{Create a map in D3}
\usage{
d3_map(shape, projection = "Mercator", stroke_col = "#fff",
  stroke_width = 0.5, width = NULL, height = NULL)
}
\arguments{
\item{shape}{A \code{sf} or \code{sp} object.}

\item{projection}{D3 projection to use.}

\item{stroke_col}{Color of shape contours.}

\item{stroke_width}{Width of shape contour in pixels.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Create a map in D3
}
\examples{
library( sf )
library( rnaturalearth )
library( r2d3maps )

# Hello world!
world <- st_as_sf(countries110)
d3_map(shape = world)

# Brazil!
brazil <- ne_states(country = "brazil", returnclass = "sf")
d3_map(shape = brazil)

# Italy!
italy <- ne_states(country = "italy", returnclass = "sf")
d3_map(shape = italy)

# Oceania!
oceania <- ne_countries(continent = "oceania", returnclass = "sf")
oceania <- sf::st_crop(oceania, xmin = 112, ymin = -56, xmax = 194, ymax = 12)
d3_map(shape = oceania) \%>\%
  add_tooltip()

}
