% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-discrete.R
\name{discrete-scale}
\alias{discrete-scale}
\alias{add_discrete_scale}
\alias{add_discrete_scale2}
\title{Add discrete scale to a map}
\usage{
add_discrete_scale(map, var, palette = "viridis", direction = 1,
  na.color = "#D8D8D8")

add_discrete_scale2(map, var, values, na.color = "#D8D8D8")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{var}{Variable to map}

\item{palette}{Color palette, you can use Viridis or Brewer color palette.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{na.color}{Color to use for missing values.}

\item{values}{Named list mapping data values to colors.
It's recommended to use Hex color code without alpha,
e.g. \code{#} followed by 6 chars \code{[0-9a-f]}.}
}
\description{
Display a discrete value on a map. \code{add_discrete_scale} is for using a color palette,
\code{add_discrete_scale2} is to attach custom colors to data levels.
}
\examples{
library( r2d3maps )
library( rnaturalearth )

# data
japan <- ne_states(country = "japan", returnclass = "sf")

# Japan's regions
d3_map(shape = japan) \%>\%
  add_discrete_scale(var = "region")

# different color palette
d3_map(shape = japan) \%>\%
  add_discrete_scale(var = "region", palette = "Set2")

# custom colors
d3_map(shape = japan) \%>\%
  add_discrete_scale2(
    var = "region",
    values = list(
      "Chugoku" = "#000080",
      "Kyushu" = "#6B8E23",
      "Shikoku" = "#DDA0DD",
      "Chubu" = "#4169E1",
      "Kinki" = "#2E8B57",
      "Hokkaido" = "#4682B4",
      "Kanto" = "#FFA07A",
      "Tohoku" = "#F08080",
      "Okinawa" = "red"
    ),
    na.color = "#000"
  )

# with legend
d3_map(shape = japan) \%>\%
  add_discrete_scale(var = "region", palette = "Set1") \%>\%
  add_legend(title = "County")

}
