% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.R
\name{gradient-scale}
\alias{gradient-scale}
\alias{add_continuous_gradient}
\alias{add_continuous_gradient2}
\title{Add gradient scale to a map}
\usage{
add_continuous_gradient(map, var, low = "#132B43", high = "#56B1F7",
  range = NULL)

add_continuous_gradient2(map, var, low = muted("red"), mid = "white",
  high = muted("blue"), range = NULL)
}
\arguments{
\item{map}{A \code{d3_map} \code{htmlwidget} object.}

\item{var}{Variable to map.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{range}{A length two vector to force range of data.}

\item{mid}{Colour for mid point.}
}
\description{
Create a two colour gradient (low-high) or
a diverging colour gradient (low-mid-high) based on a continuous variable.
}
\examples{
library( r2d3maps )
library( rnaturalearth )

# shapes
africa <- ne_countries(continent = "Africa", returnclass = "sf")

# drinking water data
data("water_africa")
wa2015 <- water_africa[water_africa$year == "2015", ]

# merge with sf object
africa <- merge(
  x = africa[, c("adm0_a3_is", "name", "geometry")],
  y = wa2015[, c("iso3", "national_at_least_basic")],
  by.x = "adm0_a3_is", by.y = "iso3"
)

africa$national_at_least_basic <- round(africa$national_at_least_basic)

# two colour gradient
d3_map(shape = africa) \%>\%
  add_continuous_gradient(
    var = "national_at_least_basic",
    range = c(0, 100)
  ) \%>\%
  add_tooltip(value = "<b>{name}</b>: {national_at_least_basic}\%") \%>\%
  add_legend(title = "Population with at least basic access", suffix = "\%") \%>\%
  add_labs(title = "Drinking water in Africa", caption = "Data: https://washdata.org/")


# three colour gradient
d3_map(shape = africa, stroke_col = "#585858") \%>\%
  add_continuous_gradient2(
    var = "national_at_least_basic",
    range = c(0, 100)
  ) \%>\%
  add_tooltip(value = "<b>{name}</b>: {national_at_least_basic}\%") \%>\%
  add_legend(title = "Population with at least basic access", suffix = "\%") \%>\%
  add_labs(title = "Drinking water in Africa", caption = "Data: https://washdata.org/")

}
