% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_fr}
\alias{pop_fr}
\title{French population by department (in 2014).}
\format{A data frame with 100 rows and 5 variables:
\describe{
  \item{code_departement}{County code}
  \item{nom_du_departement}{County name}
  \item{population_totale}{Population}
  \item{adm1_code}{Code ADM1 (for join with NaturalEarth data)}
  \item{code_hasc}{Code HASC (for join with NaturalEarth data)}
}}
\source{
data.gouv (\url{https://www.data.gouv.fr/fr/datasets/population/})
}
\usage{
pop_fr
}
\description{
French population by department (in 2014).
}
\keyword{datasets}
