% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prenoms_fr}
\alias{prenoms_fr}
\title{Most popular first names in France (1988-2016).}
\format{A data frame with 5684 rows and 7 variables:
\describe{
  \item{dpt}{County code}
  \item{sexe}{'2' for women, '1' for men}
  \item{annais}{Year of birth}
  \item{preusuel}{Most given firstname}
  \item{nombre}{Number of children given the first name}
  \item{nom_du_departement}{County name}
  \item{adm1_code}{Code ADM1 (for join with NaturalEarth data)}
}}
\source{
Insee (\url{https://www.insee.fr/fr/statistiques/2540004})
}
\usage{
prenoms_fr
}
\description{
First names most attributed to children born in France
(mainland France and Dom) between 1988 and 2016 and the
number of children by sex associated with each first name.
}
\keyword{datasets}
