% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps-package.R
\name{r2d3maps-exports}
\alias{r2d3maps-exports}
\alias{\%>\%}
\title{r2d3maps exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the r2d3maps package to avoid listing the magrittr
as Depends of r2d3maps
}
