% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps-package.R
\docType{package}
\name{r2d3maps-package}
\alias{r2d3maps-package}
\title{A wrapper to 'r2d3' to make maps from 'sf' objects thanks to 'geojsonio'}
\description{
Create interactive choropleth maps in D3
}
\author{
Victor Perrier (@dreamRs_fr)
}
