% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map_proxy.R
\name{update_continuous_gradient}
\alias{update_continuous_gradient}
\title{Update a gradient scale in Shiny}
\usage{
update_continuous_gradient(proxy, var, low = NULL, high = NULL,
  range = NULL)
}
\arguments{
\item{proxy}{A \code{d3_map_proxy} object.}

\item{var}{New var to use on the map.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{range}{A length two vector to force range of data.}
}
\description{
Update a gradient scale in Shiny
}
\examples{
\dontrun{

# todo

}
}
