% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map_proxy.R
\name{update_legend}
\alias{update_legend}
\title{Update a legend in Shiny}
\usage{
update_legend(proxy, title = "", prefix = "", suffix = "",
  d3_format = NULL)
}
\arguments{
\item{proxy}{A \code{d3_map_proxy} object.}

\item{title}{Title for the legend.}

\item{prefix}{A prefix of legend labels.}

\item{suffix}{A suffix of legend labels.}

\item{d3_format}{A string passed to \code{d3.format},
see \url{https://github.com/d3/d3-format}.
If used \code{prefix} and \code{suffix} are ignored.}
}
\description{
Update a legend in Shiny
}
\examples{
\dontrun{

# todo

}
}
