% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map_proxy.R
\name{update_tootltip}
\alias{update_tootltip}
\title{Update tooltip in Shiny}
\usage{
update_tootltip(proxy, value = "<b>{name}</b><<scale_var>>",
  as_glue = TRUE, .na = "no data")
}
\arguments{
\item{proxy}{A \code{d3_map_proxy} object.}

\item{value}{A \code{formula} or a \code{glue} string matching columns in \code{data}.}

\item{as_glue}{Use a \code{glue} string, if \code{FALSE}
you can pass a character vector as tooltip.}

\item{.na}{Value to replace NA values with (if \code{value} is a \code{glue} string).
Use \code{NULL} to don't display tooltip if there is NAs.}
}
\description{
Update tooltip in Shiny
}
\examples{
\dontrun{

if (interactive()) {



}

}
}
