% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3map.R
\name{use_r2d3map}
\alias{use_r2d3map}
\title{Use r2d3 template to create you D3 map}
\usage{
use_r2d3map(path = "my_map.R")
}
\arguments{
\item{path}{Path to a script R to create.}
}
\description{
Use r2d3 template to create you D3 map
}
\examples{
\dontrun{

# todo

}
}
