% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_continuous_scale}
\alias{add_continuous_scale}
\title{Add continuous scale to a map}
\usage{
add_continuous_scale(map, var, palette = "viridis", direction = 1,
  n_breaks = 5, range = NULL)
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{var}{Variable to map.}

\item{palette}{Color palette, you can use Viridis or Brewer color palette.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{n_breaks}{Number of breaks to cut data.}

\item{range}{Range of data, if \code{NULL} (default) range is calculated from data.
You can specify custom value, e.g. \code{c(0, 100)} to force palette to go from 0 to 100.}
}
\description{
Add continuous scale to a map
}
\examples{
\dontrun{

# todo

}
}
