% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to a map}
\usage{
add_legend(map, title = "", prefix = "", suffix = "")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{title}{Title for the legend.}

\item{prefix}{A prefix of legend labels.}

\item{suffix}{A suffix of legend labels.}
}
\value{
A \code{r2d3map} \code{htmlwidget} object.
}
\description{
Add a legend to a map
}
\examples{
\dontrun{

# todo

}
}
