% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip on a map}
\usage{
add_tooltip(map, value = "<b>{name}</b><<scale_var>>", as_glue = TRUE,
  .na = "no data")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{value}{A \code{formula} or a \code{glue} string matching columns in \code{data}.}

\item{as_glue}{Use a \code{glue} string, if \code{FALSE}
you can pass a character vector as tooltip.}

\item{.na}{Value to replace NA values with (if \code{value} is a \code{glue} string).
Use \code{NULL} to don't display tooltip if there is NAs.}
}
\value{
A \code{r2d3map} \code{htmlwidget} object.
}
\description{
Add a tooltip on a map
}
\examples{
library( r2d3maps )
library( rnaturalearth )

belgium <- ne_states(country = "belgium", returnclass = "sf")

# default
d3_map(shape = belgium) \%>\%
  add_tooltip()

# glue
d3_map(shape = belgium) \%>\%
  add_tooltip(value = "{name} ({gn_name})")

# formula
d3_map(shape = belgium) \%>\%
  add_tooltip(value = ~paste0(name, " (", gn_name, ")"))
}
