% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip on a map}
\usage{
add_tooltip(map, value = "<b>{name}</b><<scale_var>>", as_glue = TRUE,
  .na = "no data")
}
\arguments{
\item{map}{A \code{r2d3map} \code{htmlwidget} object.}

\item{value}{A \code{glue} string matching vars in \code{data}.}

\item{as_glue}{Use a \code{glue} string, if \code{FALSE}
you can pass a character vector as tooltip.}

\item{.na}{Value to replace NA values with (if \code{value} is a \code{glue} string).
Use \code{NULL} to don't display tooltip if there is NAs.}
}
\value{
A \code{r2d3map} \code{htmlwidget} object.
}
\description{
Add a tooltip on a map
}
\examples{
\dontrun{

# todo

}
}
