% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.R
\name{gradient-scale}
\alias{gradient-scale}
\alias{add_continuous_gradient}
\alias{add_continuous_gradient2}
\title{Add gradient scale to a map}
\usage{
add_continuous_gradient(map, var, low = "#132B43", high = "#56B1F7",
  range = NULL)

add_continuous_gradient2(map, var, low = muted("red"), mid = "white",
  high = muted("blue"), range = NULL)
}
\arguments{
\item{map}{A \code{d3_map} \code{htmlwidget} object.}

\item{var}{Variable to map.}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{range}{A length two vector to force range of data.}

\item{mid}{Colour for mid point.}
}
\description{
Create a two colour gradient (low-high) or
a diverging colour gradient (low-mid-high) based on a continuous variable.
}
\examples{
#todo
}
