% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3map.R
\name{r2d3map}
\alias{r2d3map}
\title{Wrapper of `r2d3` to create maps}
\usage{
r2d3map(data, script, css = "auto", dependencies = NULL,
  options = NULL, d3_version = c("5", "4", "3"), container = "svg",
  elementId = NULL, width = NULL, height = NULL,
  sizing = default_sizing(), viewer = c("internal", "external",
  "browser"))
}
\arguments{
\item{data}{An `sf` or `sp` object to convert to topojson.}

\item{script}{JavaScript file containing the D3 script.}

\item{css}{CSS file containing styles. The default value "auto" will use any CSS file
located alongside the script file with the same stem (e.g. "barplot.css" would be
used for "barplot.js") as well as any CSS file with the name "styles.css".}

\item{dependencies}{Additional HTML dependencies. These can take the form of paths to
JavaScript or CSS files, or alternatively can be fully specified dependencies created
with [htmltools::htmlDependency].}

\item{options}{Options to be passed to D3 script.}

\item{d3_version}{Major D3 version to use, the latest minor version is automatically
picked.}

\item{container}{The 'HTML' container of the D3 output.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that needs to
explicitly discover and interact with a specific widget instance.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}

\item{sizing}{Widget sizing policy (see [htmlwidgets::sizingPolicy]).}

\item{viewer}{"internal" to use the RStudio internal viewer pane for output; "external"
to display in an external RStudio window; "browser" to display in an external
browser.}
}
\description{
Wrapper of `r2d3` to create maps
}
\examples{
# todo
}
