% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3maps.R
\name{r2d3map}
\alias{r2d3map}
\title{Create a map in D3}
\usage{
r2d3map(shape, projection = "Mercator", width = NULL, height = NULL)
}
\arguments{
\item{shape}{A \code{sf} or \code{sp} object.}

\item{projection}{D3 projection to use.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Create a map in D3
}
\examples{
\dontrun{

# todo

}
}
