% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map_proxy.R
\name{update_continuous_breaks}
\alias{update_continuous_breaks}
\title{Update a continuous scale in Shiny}
\usage{
update_continuous_breaks(proxy, var, palette = NULL, direction = 1,
  n_breaks = 5, style = "pretty")
}
\arguments{
\item{proxy}{A \code{d3_map_proxy} object.}

\item{var}{New var to use on the map.}

\item{palette}{Color palette, you can use Viridis or Brewer color palette.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{n_breaks}{Number of breaks to cut data (depending on \code{style}, number of breaks can be re-computed).}

\item{style}{Style for computing breaks, see \code{\link[classInt]{classIntervals}}.}
}
\description{
Update a continuous scale in Shiny
}
\examples{
\dontrun{

# todo

}
}
