% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_map_proxy.R
\name{update_legend}
\alias{update_legend}
\title{Update a legend in Shiny}
\usage{
update_legend(proxy, title = "", prefix = "", suffix = "",
  d3_format = NULL)
}
\arguments{
\item{proxy}{A \code{d3_map_proxy} object.}

\item{title}{Title for the legend.}

\item{prefix}{A prefix of legend labels.}

\item{suffix}{A suffix of legend labels.}

\item{d3_format}{A string passed to \code{d3.format},
see \url{https://github.com/d3/d3-format}.
If used \code{prefix} and \code{suffix} are ignored.}
}
\description{
Update a legend in Shiny
}
\examples{
\dontrun{

if (interactive()) {

library(r2d3maps)
library(shiny)

# data about Paris
data("paris")


# app
ui <- fluidPage(
  fluidRow(
    column(
      width = 8, offset = 2,
      tags$h2("Proxy for continuous breaks scale"),
      d3Output(outputId = "mymap"),
      selectInput(
        inputId = "var", label = "Variable:",
        choices = grep(pattern = "AGE", x = names(paris), value = TRUE)
      )
    )
  )
)

server <- function(input, output, session) {

  output$mymap <- renderD3({
    d3_map(shape = paris) \%>\%
      add_continuous_gradient(var = "AGE_00", low = "#FEE0D2", high = "#CB181D") \%>\%
      add_legend(d3_format = ".2s")
  })

  observeEvent(input$var, {
    d3_map_proxy(shinyId = "mymap", data = paris) \%>\%
      update_continuous_gradient(var = input$var) \%>\%
      update_legend(title = tolower(gsub(
        patter = "_", replacement = " ", x = input$var
      )), d3_format = ".1s")
  }, ignoreInit = TRUE)

}

shinyApp(ui, server)

}

}
}
